requirements.txt
pyserial>=3.5
Pillow>=9.5
Flask>=2.2
requests>=2.31
qrcode[pil]>=7.4
fastapi>=0.110
uvicorn[standard]>=0.29
pyzbar>=0.1
PyYAML>=6.0
openai>=1.14
